# [match] add caching layer to significantly improve performance by up to 100x