<!-- Thanks for contributing to _fastlane_! Before you submit your pull request, please make sure to check the following boxes by putting an x in the [ ] (don't: [x ], [ x], do: [x]) -->

### Checklist

- [x] I've run `bundle exec rspec` from the root directory to see all new and existing tests pass
- [x] I've followed the _fastlane_ code style and run `bundle exec rubocop -a` to ensure the code style is valid
- [x] I see several green `ci/circleci` builds in the "All checks have passed" section of my PR ([connect CircleCI to GitHub](https://support.circleci.com/hc/en-us/articles/360008097173-Why-aren-t-pull-requests-triggering-jobs-on-my-organization-) if not)
- [x] I've read the [Contribution Guidelines](https://github.com/fastlane/fastlane/blob/master/CONTRIBUTING.md)
- [x] I've updated the documentation if necessary.
- [x] I've added or updated relevant unit tests.

Hi.

## Problem
We have 112 provisioning profiles. And running match not in `readonly` mode takes a lot of time for us.
I've made a Faraday middleware to count requests, fetch time, and the amount of passed data.
I saw many redundant fetches: profiles, certificates, bundle IDs, and devices fetched for each `app_identifier` and some of them several times.

tl;dr; The cached version of the match installs all 112 profiles with cert and device checks as fast as the current version manages to check and install just **one** development profile with the same parameters.

## Description

Introduce cache and fetcher for match. We must save the state of profiles, certificates, bundle IDs, and devices. I've looked at what information changes when the match is executed and what is static.
* Bundle IDs: static. match doesn't modify bundle identifiers.
* Devices: static. match doesn't modify devices.
* Provisioning: static. match just checks if the current provisioning is valid. After generating a new one, we don't need to update or reset the cache with a new provisioning profile because it's unused for the next app_identifier.
* Certificates: dynamic but close to static. When the match generates a new certificate, it generates it only once for all app_identifiers, and resetting the cache once is not a big problem.
With this, I see that caching will not bring a lot of headaches, and there will be no cache inconsistency. And introducing cache on the instance level will probably bring more good than bad.

Another good thing is that the match is executed with strict parameters: type and platform. And we can write better filters for fetching certificates and profiles.

## Results (x100 😅)

| | Before | After | % |
|-|-|-|-|
| calls count | 1,894 | 21 | 0.011 |
| duration in seconds | 4,957.58 | 30.57 | 0.006 |
| response size | 2,702,427,979 | 16,945,609 | 0.006 |

### Stats per url
Before:
<details>

```json
{
  "https://api.appstoreconnect.apple.com/v1/bundleIds": {
    "duration": 85.95213599999997,
    "size": 285596,
    "calls_count": 112
  },
  "https://api.appstoreconnect.apple.com/v1/certificates": {
    "duration": 282.9322100000001,
    "size": 12128325,
    "calls_count": 244
  },
  "https://api.appstoreconnect.apple.com/v1/profiles": {
    "duration": 4370.683801999995,
    "size": 2687726458,
    "calls_count": 1258
  },
  "https://api.appstoreconnect.apple.com/v1/devices": {
    "duration": 218.01276800000005,
    "size": 2287600,
    "calls_count": 280
  }
}
```

</details>

After:
<details>

```json
{
  "https://api.appstoreconnect.apple.com/v1/bundleIds": {
    "duration": 9.408271000000001,
    "size": 199607,
    "calls_count": 9
  },
  "https://api.appstoreconnect.apple.com/v1/certificates": {
    "duration": 5.092039,
    "size": 378884,
    "calls_count": 5
  },
  "https://api.appstoreconnect.apple.com/v1/profiles": {
    "duration": 14.223327999999999,
    "size": 16328383,
    "calls_count": 5
  },
  "https://api.appstoreconnect.apple.com/v1/devices": {
    "duration": 1.850342,
    "size": 38735,
    "calls_count": 2
  }
}
```
</details>

## Solution

This PR.

To ease the review, I can split it into 5 smaller PRs:

1. Move `device_count_different?` and `certificate_count_different?` to a new file.
2. Slightly refactor sigh: move profile_type and certificate_types map functions to a module.
3. Create a Match::Portal::Fetcher to have a controlled entry point for API requests.
4. Add Match::Portal::Cache.
5. Inject cached data into the sigh when creating profiles.

As for me, making separate PRs into the main branch makes almost no sense. I'm a fan of a feature branch but not a maintainer 😅. If you are ready for one biggie, please let me know.

Please tell me if my solution can be merged into the fastlane, and what strategy of PRs you would like to use.

PS. I want to thank @PaulTaykalo for his #21016 PR, which I included in this one. Paul, I hope you don't mind 😉.