@nekrich 2023-12-12
@rogerluan ,
> 1. Is the cache persistent (e.g. on disk)? If so, what's the persistency scope (e.g. is it valid throughout different _fastlane_ action calls of a single lane, or different _fastlane_ lane calls?
> 2. What is the cache bursting logic?

1. It's in-memory cache.
2. It works only in match. match action is responsible for its own cache.
3. There is an option to pass cached objects to the sigh from match.
4. Cache lifetime is a single match run. 
  It has two internal invalidation points: 
    * The whole certificates cache is reset on cert generation.
    * "Forgetting" a single provision after a new one was created. Which now is redundant after fixing a @lacostej [comment](https://github.com/fastlane/fastlane/pull/21694/commits/39b784935b9129e43b9e19ca362d6d88198a9c1d#r1417089753).
  
    No long-term persistence lane or a whole fastlane run. And no headache on what to do if the lane runs 30 mins, and we have another Fastlane run on another computer which made changes to certs/profiles.
    
    It doesn't fetch irrelevant info and utilizes available filters in the API — filter cert by types, fetch only active devices, only listed bundle ids.