@nekrich 2024-01-03
> @nekrich I'm experiencing an issue with the error message "cached_profiles parameter must be a non-empty array of Spaceship::ConnectAPI::Profile" after upgrading to fastlane version 2.218.0. I had to revert to version 2.217.0 to resolve the issue.
> 
> Any suggestions on how to fix it?

@thanhcuong1990 , I'll look into in in a couple of hours. Can you please tell what auth type you use: Keys or login and password?