@thanhcuong1990 2024-01-03
> @thanhcuong1990 , I'll look into in in a couple of hours. Can you please tell what auth type you use: Keys or login and password?

I used `FASTLANE_APPLE_APPLICATION_SPECIFIC_PASSWORD` for auth.