@viktor-savchik-idf 2024-01-03
The same issue. I'm using *.p8 key
```
[11:57:22]: Error setting value '[]' for option 'cached_profiles'
+-----------------------------------------+
|              Lane Context               |
+------------------+----------------------+
| DEFAULT_PLATFORM | ios                  |
| PLATFORM_NAME    | ios                  |
| LANE_NAME        | ios certificates_dev |
+------------------+----------------------+
[11:57:22]: cached_profiles parameter must be a non-empty array of Spaceship::ConnectAPI::Profile

+---------------------------------------------------------------------------+
|                             fastlane summary                              |
+------+------------------------------------------------------+-------------+
| Step | Action                                               | Time (in s) |
+------+------------------------------------------------------+-------------+
| 1    | default_platform                                     | 0           |
| 2    | Switch to ios app_store_connect_api_key_creator lane | 0           |
| 3    | app_store_connect_api_key                            | 0           |
| 4    | register_devices                                     | 1           |
| 💥   | match                                                | 9           |
+------+------------------------------------------------------+-------------+

+-----------------------------------------------------------+
|                 Plugin updates available                  |
+---------------------------+--------------+----------------+
| Plugin                    | Your Version | Latest Version |
+---------------------------+--------------+----------------+
| firebase_app_distribution | 0.7.4        | 0.8.0          |
+---------------------------+--------------+----------------+
[11:57:22]: To update all plugins, just run
[11:57:22]: $ bundle exec fastlane update_plugins

[11:57:22]: fastlane finished with errors

[!] cached_profiles parameter must be a non-empty array of Spaceship::ConnectAPI::Profile
 ✘ viktor_savchik@onexoer  ~/path/to/project/ios   master  bundle install
```